update cart ui and fetch counter details
