+++ src/pages/MerchantPanel.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const MerchantPanel = () => {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default MerchantPanel
